move object creation to factory.c to make testing possible
