+++ include/express/factory.h
@@ -1,0 +1,6 @@
+#ifndef __FACTORY_H_
+#define __FACTORY_H_
+
+void FACTORYinitialize();
+
+#endif /* __FACTORY_H_ */
