+++ src/express/type.c
@@ -126,38 +126,6 @@ This module implements the type abstraction.  It is
 #include <sc_memmgr.h>
 #include "express/type.h"
 
-/* Very commonly-used read-only types */
-/* non-constant versions probably aren't necessary? */
-Type Type_Bad;
-Type Type_Unknown;
-Type Type_Dont_Care;
-Type Type_Runtime; /* indicates that this object can't be */
-/* calculated now but must be deferred */
-/* til (the mythical) runtime */
-Type Type_Binary;
-Type Type_Boolean;
-Type Type_Enumeration;
-Type Type_Expression;
-Type Type_Aggregate;
-Type Type_Repeat;
-Type Type_Integer;
-Type Type_Number;
-Type Type_Real;
-Type Type_String;
-Type Type_String_Encoded;
-Type Type_Logical;
-Type Type_Set;
-Type Type_Attribute;
-Type Type_Entity;
-Type Type_Funcall;
-Type Type_Generic;
-Type Type_Identifier;
-Type Type_Oneof;
-Type Type_Query;
-Type Type_Self;
-Type Type_Set_Of_String;
-Type Type_Set_Of_Generic;
-Type Type_Bag_Of_Generic;
 
 Error ERROR_corrupted_type = ERROR_none;
 
@@ -258,104 +226,6 @@ return( false );
 
 /** Initialize the Type module */
 void TYPEinitialize() {
-    /* Very commonly-used read-only types */
-    Type_Unknown = TYPEcreate( unknown_ );
-    Type_Dont_Care = TYPEcreate( special_ );
-    Type_Bad = TYPEcreate( special_ );
-    Type_Runtime = TYPEcreate( runtime_ );
-
-    Type_Enumeration = TYPEcreate( enumeration_ );
-    Type_Enumeration->u.type->body->flags.shared = 1;
-    resolved_all( Type_Enumeration );
-
-    Type_Expression = TYPEcreate( op_ );
-    Type_Expression->u.type->body->flags.shared = 1;
-
-    Type_Aggregate = TYPEcreate( aggregate_ );
-    Type_Aggregate->u.type->body->flags.shared = 1;
-    Type_Aggregate->u.type->body->base = Type_Runtime;
-
-    Type_Integer = TYPEcreate( integer_ );
-    Type_Integer->u.type->body->flags.shared = 1;
-    resolved_all( Type_Integer );
-
-    Type_Real = TYPEcreate( real_ );
-    Type_Real->u.type->body->flags.shared = 1;
-    resolved_all( Type_Real );
-
-    Type_Number = TYPEcreate( number_ );
-    Type_Number->u.type->body->flags.shared = 1;
-    resolved_all( Type_Number );
-
-    Type_String = TYPEcreate( string_ );
-    Type_String->u.type->body->flags.shared = 1;
-    resolved_all( Type_String );
-
-    Type_String_Encoded = TYPEcreate( string_ );
-    Type_String_Encoded->u.type->body->flags.shared = 1;
-    Type_String_Encoded->u.type->body->flags.encoded = 1;
-    resolved_all( Type_String );
-
-    Type_Logical = TYPEcreate( logical_ );
-    Type_Logical->u.type->body->flags.shared = 1;
-    resolved_all( Type_Logical );
-
-    Type_Binary = TYPEcreate( binary_ );
-    Type_Binary->u.type->body->flags.shared = 1;
-    resolved_all( Type_Binary );
-
-    Type_Number = TYPEcreate( number_ );
-    Type_Number->u.type->body->flags.shared = 1;
-    resolved_all( Type_Number );
-
-    Type_Boolean = TYPEcreate( boolean_ );
-    Type_Boolean->u.type->body->flags.shared = 1;
-    resolved_all( Type_Boolean );
-
-    Type_Generic = TYPEcreate( generic_ );
-    Type_Generic->u.type->body->flags.shared = 1;
-    resolved_all( Type_Generic );
-
-    Type_Set_Of_String = TYPEcreate( set_ );
-    Type_Set_Of_String->u.type->body->flags.shared = 1;
-    Type_Set_Of_String->u.type->body->base = Type_String;
-
-    Type_Set_Of_Generic = TYPEcreate( set_ );
-    Type_Set_Of_Generic->u.type->body->flags.shared = 1;
-    Type_Set_Of_Generic->u.type->body->base = Type_Generic;
-
-    Type_Bag_Of_Generic = TYPEcreate( bag_ );
-    Type_Bag_Of_Generic->u.type->body->flags.shared = 1;
-    Type_Bag_Of_Generic->u.type->body->base = Type_Generic;
-
-    Type_Attribute = TYPEcreate( attribute_ );
-    Type_Attribute->u.type->body->flags.shared = 1;
-
-    Type_Entity = TYPEcreate( entity_ );
-    Type_Entity->u.type->body->flags.shared = 1;
-
-    Type_Funcall = TYPEcreate( funcall_ );
-    Type_Funcall->u.type->body->flags.shared = 1;
-
-    Type_Generic = TYPEcreate( generic_ );
-    Type_Generic->u.type->body->flags.shared = 1;
-
-    Type_Identifier = TYPEcreate( identifier_ );
-    Type_Identifier->u.type->body->flags.shared = 1;
-
-    Type_Repeat = TYPEcreate( integer_ );
-    Type_Repeat->u.type->body->flags.shared = 1;
-    Type_Repeat->u.type->body->flags.repeat = 1;
-
-    Type_Oneof = TYPEcreate( oneof_ );
-    Type_Oneof->u.type->body->flags.shared = 1;
-
-    Type_Query = TYPEcreate( query_ );
-    Type_Query->u.type->body->flags.shared = 1;
-
-    Type_Self = TYPEcreate( self_ );
-    Type_Self->u.type->body->flags.shared = 1;
-
     ERROR_corrupted_type =
         ERRORcreate( "Corrupted type in %s", SEVERITY_DUMP );
